requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
asyncpg==0.29.0
sqlalchemy==2.0.32
alembic==1.13.2
psycopg2-binary==2.9.10

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.27.2

# AWS S3 & Object Storage
boto3==1.34.0
botocore==1.34.0

# Caching & Performance
redis==5.0.1

# Validation & Serialization
pydantic==2.8.2
pydantic-settings==2.4.0
email-validator==2.1.0

# Search & Text Processing
asyncpg-stubs==0.29.1

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Rate Limiting & Security
slowapi==0.1.9

# File Processing
aiofiles==23.2.1
python-magic==0.4.27

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.27.2
factory-boy==3.3.0

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# YAML Processing
PyYAML==6.0.1
pyyaml-include==1.3.1

# Date/Time
python-dateutil==2.8.2

# Environment Variables
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6

# JSON Schema Validation
jsonschema==4.23.0

# Package Version Handling
semantic-version==2.10.0
packaging==23.2

# Background Tasks
celery==5.3.4
flower==2.0.1

# Metrics & Health Checks
prometheus-client==0.19.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# API specification and validation
pyyaml==6.0.1
requests==2.32.3 